Decouple activity logging from writes; back award totals with a counter column

+++ dundie-awards-service/src/main/java/com/ninjaone/dundie_awards/dto/LeaderboardEntry.java
@@ -4,5 +4,5 @@ public record LeaderboardEntry(
         long recipientId,
         String recipientName,
         String organizationName,
-        long awardCount) {
+        int awardCount) {
 }
